+++ packages/plugin-image/web/src/image-component.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Loader, ErrorMsgWithIcon } from 'wix-rich-content-editor-common';
+import { Loader, MediaItemErrorMsg } from 'wix-rich-content-editor-common';
 import ImageViewer from './image-viewer';
 import { DEFAULTS } from './consts';
 import { sizeClassName, alignmentClassName } from './classNameStrategies';
@@ -53,8 +53,7 @@ class ImageComponent extends React.Component {
 
   resetLoadingState = error => {
     const dataUrl = error ? this.state.dataUrl || EMPTY_SMALL_PLACEHOLDER : null;
-    const errorMsg = error?.msg;
-    this.setState({ isLoading: false, dataUrl, errorMsg });
+    this.setState({ isLoading: false, dataUrl, error });
     this.props.store.update('componentState', { isLoading: false, userSelectedFiles: null });
   };
 
@@ -71,7 +70,7 @@ class ImageComponent extends React.Component {
     const file = files[0];
     if (file) {
       this.fileReader(file).then(dataUrl => {
-        this.setState({ isLoading: true, errorMsg: false, dataUrl });
+        this.setState({ isLoading: true, error: false, dataUrl });
         this.uploadFile(file);
       });
     }
@@ -87,10 +86,6 @@ class ImageComponent extends React.Component {
   }
 
   handleFilesAdded = ({ data, error }) => {
-    if (error) {
-      this.resetLoadingState(error);
-      return;
-    }
     const imageData = data.length ? data[0] : data;
     const config = { ...this.props.componentData.config };
     if (!config.alignment) {
@@ -99,9 +94,10 @@ class ImageComponent extends React.Component {
     const componentData = {
       config,
       src: imageData,
+      error,
     };
     this.props.store.update('componentData', componentData, this.props.block.getKey());
-    this.resetLoadingState();
+    this.resetLoadingState(error);
   };
 
   handleMetadataChange = newMetadata => {
@@ -140,9 +136,10 @@ class ImageComponent extends React.Component {
       getInPluginEditingMode,
       setInPluginEditingMode,
       setComponentUrl,
+      t,
     } = this.props;
 
-    const { errorMsg } = this.state;
+    const { error } = componentData;
     return (
       <>
         <ImageViewer
@@ -165,7 +162,7 @@ class ImageComponent extends React.Component {
           setComponentUrl={setComponentUrl}
         />
         {(this.state.isLoading || componentData?.loading) && this.renderLoader()}
-        {errorMsg && <ErrorMsgWithIcon errorMsg={errorMsg} />}
+        {error && <MediaItemErrorMsg error={error} t={t} />}
       </>
     );
   }